refactor(competition): extract active board row into GameRow (identical render)
Phase 2 §A groundwork. The per-game board row (formerly the in-file SessionRow)
moves verbatim into its own exported component, src/components/competition/
GameRow.tsx, and SessionBreakdown points at it. Pure extraction — the JSX is
byte-identical, so the active board renders exactly as before (verified by eye:
all 11 BBMI Cup session rows unchanged — name, Live/Upcoming badge, status line,
chevron; Poker still chevron-less with no route).

The row-only helpers move with it (gameHref/GAME_ROUTES, rowStateOf/RowState,
RowBadge, YoursBadge) plus the shared fmtPts (re-exported), so the only runtime
import edge is CompetitionLeaderboard → GameRow — no cycle. The LB* row types are
exported from CompetitionLeaderboard and type-imported by GameRow (erased, no
cycle).

EarlyState still has its own inline pre-game row; it folds into this same
GameRow behind a lifecycle-state prop next (Commit 2), then Phase 3 layers the
§A state machine onto the unified component.
